gateway/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
redis[hiredis]>=5.0.1
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-exporter-otlp-proto-grpc>=1.22.0
pydantic>=2.5.0
psutil>=5.9.8
PyJWT>=2.8.0
mlflow>=2.10.0
